# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lelouch Britannia Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AttackStatus = (typeof AttackStatus)[keyof typeof AttackStatus];

export const AttackStatus = {
  running: "running",
  stopped: "stopped",
  finished: "finished",
  error: "error",
} as const;

export interface Attack {
  id: number;
  target: string;
  port: number;
  method: string;
  duration: number;
  threads: number;
  status: AttackStatus;
  packetsSent: number | null;
  bytesSent: number | null;
  webhookUrl: string | null;
  startedAt: string;
  stoppedAt: string | null;
  createdAt: string;
}

export interface CreateAttackBody {
  target: string;
  port: number;
  method: string;
  duration: number;
  threads: number;
  webhookUrl?: string | null;
}

export interface MethodCount {
  method: string;
  count: number;
}

export interface AttackStats {
  totalAttacks: number;
  runningAttacks: number;
  totalPacketsSent: number;
  totalBytesSent: number;
  attacksByMethod: MethodCount[];
  recentAttacks: Attack[];
}

export type MethodLayer = (typeof MethodLayer)[keyof typeof MethodLayer];

export const MethodLayer = {
  L3: "L3",
  L4: "L4",
  L7: "L7",
} as const;

export type MethodProtocol =
  (typeof MethodProtocol)[keyof typeof MethodProtocol];

export const MethodProtocol = {
  UDP: "UDP",
  TCP: "TCP",
  HTTP: "HTTP",
  ICMP: "ICMP",
} as const;

export interface Method {
  id: string;
  name: string;
  layer: MethodLayer;
  protocol: MethodProtocol;
  description: string;
}

export interface CheckSiteBody {
  url: string;
}

export interface CheckSiteResult {
  up: boolean;
  status: number;
  statusText: string;
  responseTime: number;
  error: string | null;
}

export interface AnalyzeBody {
  url: string;
}

export type MethodRecommendationTier =
  (typeof MethodRecommendationTier)[keyof typeof MethodRecommendationTier];

export const MethodRecommendationTier = {
  S: "S",
  A: "A",
  B: "B",
  C: "C",
  D: "D",
} as const;

export interface MethodRecommendation {
  method: string;
  name: string;
  score: number;
  reason: string;
  suggestedThreads: number;
  suggestedDuration: number;
  protocol: string;
  amplification: number;
  tier: MethodRecommendationTier;
}

export interface AnalyzeResult {
  target: string;
  ip: string | null;
  isIP: boolean;
  httpAvailable: boolean;
  httpsAvailable: boolean;
  responseTimeMs: number;
  serverHeader: string;
  isCDN: boolean;
  cdnProvider: string;
  openPorts: number[];
  recommendations: MethodRecommendation[];
}

export interface InfinityLoginBody {
  username: string;
  password: string;
}

export type InfinityUserRole =
  (typeof InfinityUserRole)[keyof typeof InfinityUserRole];

export const InfinityUserRole = {
  admin: "admin",
  user: "user",
} as const;

export interface InfinityUser {
  username: string;
  role: InfinityUserRole;
  createdAt: string;
  lastLoginAt: string | null;
}

export interface InfinitySession {
  token: string;
  user: InfinityUser;
}

export type InfinityCreateUserBodyRole =
  (typeof InfinityCreateUserBodyRole)[keyof typeof InfinityCreateUserBodyRole];

export const InfinityCreateUserBodyRole = {
  admin: "admin",
  user: "user",
} as const;

export interface InfinityCreateUserBody {
  username: string;
  password: string;
  role: InfinityCreateUserBodyRole;
}

export interface InfinityTypeCount {
  tipo: string;
  count: number;
}

export interface InfinityConsulta {
  id: number;
  tipo: string;
  query: string;
  username: string;
  success: boolean;
  createdAt: string;
}

export interface InfinityOverview {
  totalConsultas: number;
  consultasHoje: number;
  consultasSemana: number;
  usuariosAtivos: number;
  consultasPorTipo: InfinityTypeCount[];
  recentes: InfinityConsulta[];
}

export interface InfinityCpfBody {
  cpf: string;
}

export interface InfinityCnpjBody {
  cnpj: string;
}

export interface InfinityTelefoneBody {
  telefone: string;
}

export type InfinityConsultaResultData = { [key: string]: unknown };

export interface InfinityConsultaResult {
  success: boolean;
  tipo: string;
  query: string;
  data: InfinityConsultaResultData;
  error: string | null;
}

export type InfinityListConsultasParams = {
  limit?: number;
};