# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Lelouch Britannia Panel API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: attacks
  - name: methods
  - name: check
  - name: analyze
  - name: infinity_auth
  - name: infinity_users
  - name: infinity_consultas
  - name: infinity_overview
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /attacks:
    get:
      operationId: listAttacks
      tags: [attacks]
      summary: List all attacks
      responses:
        "200":
          description: List of attacks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attack"
    post:
      operationId: createAttack
      tags: [attacks]
      summary: Create and launch an attack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttackBody"
      responses:
        "201":
          description: Attack created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attack"
        "400":
          description: Invalid input

  /attacks/stats:
    get:
      operationId: getAttackStats
      tags: [attacks]
      summary: Aggregate stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttackStats"

  /attacks/{id}:
    get:
      operationId: getAttack
      tags: [attacks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attack
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attack"
        "404":
          description: Not found
    delete:
      operationId: deleteAttack
      tags: [attacks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /attacks/{id}/stop:
    post:
      operationId: stopAttack
      tags: [attacks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attack"
        "404":
          description: Not found

  /methods:
    get:
      operationId: listMethods
      tags: [methods]
      summary: List attack methods
      responses:
        "200":
          description: Methods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Method"

  /check:
    post:
      operationId: checkSite
      tags: [check]
      summary: Check site status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckSiteBody"
      responses:
        "200":
          description: Check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckSiteResult"

  /analyze:
    post:
      operationId: analyzeTarget
      tags: [analyze]
      summary: Analyze a target and recommend the best attack methods
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeBody"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeResult"
        "400":
          description: Invalid input

  # ─────────────────────────── Infinity Search ───────────────────────────
  /infinity/login:
    post:
      operationId: infinityLogin
      tags: [infinity_auth]
      summary: Login to Infinity Search panel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InfinityLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfinitySession"
        "401":
          description: Invalid credentials

  /infinity/logout:
    post:
      operationId: infinityLogout
      tags: [infinity_auth]
      summary: Logout
      responses:
        "204":
          description: Logged out

  /infinity/me:
    get:
      operationId: infinityMe
      tags: [infinity_auth]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfinityUser"
        "401":
          description: Not authenticated

  /infinity/users:
    get:
      operationId: infinityListUsers
      tags: [infinity_users]
      summary: List all panel users (admin only)
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InfinityUser"
        "403":
          description: Forbidden
    post:
      operationId: infinityCreateUser
      tags: [infinity_users]
      summary: Create a new panel user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InfinityCreateUserBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfinityUser"

  /infinity/users/{username}:
    delete:
      operationId: infinityDeleteUser
      tags: [infinity_users]
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /infinity/overview:
    get:
      operationId: infinityOverview
      tags: [infinity_overview]
      summary: Dashboard overview stats
      responses:
        "200":
          description: Overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfinityOverview"

  /infinity/consultas:
    get:
      operationId: infinityListConsultas
      tags: [infinity_consultas]
      summary: List recent consultas
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Consulta history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InfinityConsulta"

  /infinity/consultas/cpf:
    post:
      operationId: infinityConsultaCpf
      tags: [infinity_consultas]
      summary: Consulta de CPF
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InfinityCpfBody"
      responses:
        "200":
          description: Resultado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfinityConsultaResult"

  /infinity/consultas/cnpj:
    post:
      operationId: infinityConsultaCnpj
      tags: [infinity_consultas]
      summary: Consulta de CNPJ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InfinityCnpjBody"
      responses:
        "200":
          description: Resultado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfinityConsultaResult"

  /infinity/consultas/telefone:
    post:
      operationId: infinityConsultaTelefone
      tags: [infinity_consultas]
      summary: Consulta de telefone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InfinityTelefoneBody"
      responses:
        "200":
          description: Resultado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfinityConsultaResult"

  /infinity/consultas/sipni:
    post:
      operationId: infinityConsultaSipni
      tags: [infinity_consultas]
      summary: Consulta SIPNI (vacinação) por CPF
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InfinityCpfBody"
      responses:
        "200":
          description: Resultado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfinityConsultaResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Attack:
      type: object
      properties:
        id:
          type: integer
        target:
          type: string
        port:
          type: integer
        method:
          type: string
        duration:
          type: integer
        threads:
          type: integer
        status:
          type: string
          enum: [running, stopped, finished, error]
        packetsSent:
          anyOf:
            - type: integer
            - type: "null"
        bytesSent:
          anyOf:
            - type: integer
            - type: "null"
        webhookUrl:
          anyOf:
            - type: string
            - type: "null"
        startedAt:
          type: string
          format: date-time
        stoppedAt:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
        createdAt:
          type: string
          format: date-time
      required: [id,target,port,method,duration,threads,status,packetsSent,bytesSent,webhookUrl,startedAt,stoppedAt,createdAt]

    CreateAttackBody:
      type: object
      properties:
        target:
          type: string
        port:
          type: integer
        method:
          type: string
        duration:
          type: integer
        threads:
          type: integer
        webhookUrl:
          anyOf:
            - type: string
            - type: "null"
      required: [target,port,method,duration,threads]

    AttackStats:
      type: object
      properties:
        totalAttacks:
          type: integer
        runningAttacks:
          type: integer
        totalPacketsSent:
          type: integer
        totalBytesSent:
          type: integer
        attacksByMethod:
          type: array
          items:
            $ref: "#/components/schemas/MethodCount"
        recentAttacks:
          type: array
          items:
            $ref: "#/components/schemas/Attack"
      required: [totalAttacks,runningAttacks,totalPacketsSent,totalBytesSent,attacksByMethod,recentAttacks]

    MethodCount:
      type: object
      properties:
        method:
          type: string
        count:
          type: integer
      required: [method, count]

    Method:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        layer:
          type: string
          enum: [L3, L4, L7]
        protocol:
          type: string
          enum: [UDP, TCP, HTTP, ICMP]
        description:
          type: string
      required: [id, name, layer, protocol, description]

    CheckSiteBody:
      type: object
      properties:
        url:
          type: string
      required: [url]

    CheckSiteResult:
      type: object
      properties:
        up:
          type: boolean
        status:
          type: integer
        statusText:
          type: string
        responseTime:
          type: integer
        error:
          anyOf:
            - type: string
            - type: "null"
      required: [up, status, statusText, responseTime, error]

    AnalyzeBody:
      type: object
      properties:
        url:
          type: string
      required: [url]

    AnalyzeResult:
      type: object
      properties:
        target:
          type: string
        ip:
          anyOf:
            - type: string
            - type: "null"
        isIP:
          type: boolean
        httpAvailable:
          type: boolean
        httpsAvailable:
          type: boolean
        responseTimeMs:
          type: integer
        serverHeader:
          type: string
        isCDN:
          type: boolean
        cdnProvider:
          type: string
        openPorts:
          type: array
          items:
            type: integer
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/MethodRecommendation"
      required: [target,ip,isIP,httpAvailable,httpsAvailable,responseTimeMs,serverHeader,isCDN,cdnProvider,openPorts,recommendations]

    InfinityLoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]

    InfinityUser:
      type: object
      properties:
        username:
          type: string
        role:
          type: string
          enum: [admin, user]
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
      required: [username, role, createdAt, lastLoginAt]

    InfinitySession:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/InfinityUser"
      required: [token, user]

    InfinityCreateUserBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, user]
      required: [username, password, role]

    InfinityOverview:
      type: object
      properties:
        totalConsultas:
          type: integer
        consultasHoje:
          type: integer
        consultasSemana:
          type: integer
        usuariosAtivos:
          type: integer
        consultasPorTipo:
          type: array
          items:
            $ref: "#/components/schemas/InfinityTypeCount"
        recentes:
          type: array
          items:
            $ref: "#/components/schemas/InfinityConsulta"
      required: [totalConsultas, consultasHoje, consultasSemana, usuariosAtivos, consultasPorTipo, recentes]

    InfinityTypeCount:
      type: object
      properties:
        tipo:
          type: string
        count:
          type: integer
      required: [tipo, count]

    InfinityConsulta:
      type: object
      properties:
        id:
          type: integer
        tipo:
          type: string
        query:
          type: string
        username:
          type: string
        success:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, tipo, query, username, success, createdAt]

    InfinityCpfBody:
      type: object
      properties:
        cpf:
          type: string
      required: [cpf]

    InfinityCnpjBody:
      type: object
      properties:
        cnpj:
          type: string
      required: [cnpj]

    InfinityTelefoneBody:
      type: object
      properties:
        telefone:
          type: string
      required: [telefone]

    InfinityConsultaResult:
      type: object
      properties:
        success:
          type: boolean
        tipo:
          type: string
        query:
          type: string
        data:
          type: object
          additionalProperties: true
        error:
          anyOf:
            - type: string
            - type: "null"
      required: [success, tipo, query, data, error]

    MethodRecommendation:
      type: object
      properties:
        method:
          type: string
        name:
          type: string
        score:
          type: integer
        reason:
          type: string
        suggestedThreads:
          type: integer
        suggestedDuration:
          type: integer
        protocol:
          type: string
        amplification:
          type: number
        tier:
          type: string
          enum: [S, A, B, C, D]
      required: [method, name, score, reason, suggestedThreads, suggestedDuration, protocol, amplification, tier]