# lib/api-zod/src/generated/types/analyzeResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lelouch Britannia Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { MethodRecommendation } from "./methodRecommendation";

export interface AnalyzeResult {
  target: string;
  ip: string | null;
  isIP: boolean;
  httpAvailable: boolean;
  httpsAvailable: boolean;
  responseTimeMs: number;
  serverHeader: string;
  isCDN: boolean;
  cdnProvider: string;
  openPorts: number[];
  recommendations: MethodRecommendation[];
}