# lib/api-zod/src/generated/types/attackStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lelouch Britannia Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { Attack } from "./attack";
import type { MethodCount } from "./methodCount";

export interface AttackStats {
  totalAttacks: number;
  runningAttacks: number;
  totalPacketsSent: number;
  totalBytesSent: number;
  attacksByMethod: MethodCount[];
  recentAttacks: Attack[];
}